Conexion del get con la base de datos. Se inicializa con user_1 cambios de int a long generacion de id no secuencial

+++ ApplicationServer/src/api/handler/Handler.cpp
@@ -82,14 +82,14 @@ string Handler::getSubUrl(string url) {
     return subUrl;
 }
 
-int Handler::getUserIdFromUrl(string url) {
+long Handler::getUserIdFromUrl(string url) {
     vector<string> parsedUrl = this->parseUrl(url);
     if (parsedUrl.size() != 1) {
         throw InvalidRequestException("Cannot get user id from url.");
     }
     string userIdAsString = parsedUrl[0];
     try {
-        int userId = stoi(userIdAsString);
+        long userId = stol(userIdAsString);
         return userId;
     } catch (invalid_argument e) {
         throw InvalidRequestException("Not a numeric id");
